Improvement in the smooth-abs activation

+++ unittest/factory/solver.cpp
@@ -56,11 +56,11 @@ boost::shared_ptr<crocoddyl::SolverAbstract> SolverFactory::create(SolverTypes::
   boost::shared_ptr<crocoddyl::ActionModelAbstract> model = ActionModelFactory().create(action_type);
   boost::shared_ptr<crocoddyl::ActionModelAbstract> model2 = ActionModelFactory().create(action_type, true);
   std::vector<boost::shared_ptr<crocoddyl::ActionModelAbstract> > running_models;
-  const int halfway = int(T / 2);
-  for (int i = 0; i < halfway; ++i) {
+  const size_t halfway = T / 2;
+  for (size_t i = 0; i < halfway; ++i) {
     running_models.push_back(model);
   }
-  for (int i = 0; i < T - halfway; ++i) {
+  for (size_t i = 0; i < T - halfway; ++i) {
     running_models.push_back(model2);
   }
